feat: cache-first logic
