+++ app/src/main/res/drawable/app_ic_dates_cloud_fill.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="20dp"
+    android:viewportWidth="18"
+    android:viewportHeight="20">
+    <path
+        android:pathData="M16,2H15V0H13V2H5V0H3V2H2C0.89,2 0.01,2.9 0.01,4L0,18C0,19.1 0.89,20 2,20H16C17.1,20 18,19.1 18,18V4C18,2.9 17.1,2 16,2ZM16,18H2V8H16V18ZM6,12H4V10H6V12ZM10,12H8V10H10V12ZM14,12H12V10H14V12ZM6,16H4V14H6V16ZM10,16H8V14H10V16ZM14,16H12V14H14V16Z"
+        android:fillColor="#3C68FF" />
+</vector>
