+++ app/src/main/res/drawable/app_ic_dates_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/app_ic_dates_cloud_fill" android:state_checked="true" />
+    <item android:drawable="@drawable/app_ic_dates_cloud_outline" />
+</selector>
