+++ app/src/main/res/values/main_manu_tab_ids.xml
@@ -3,5 +3,6 @@
     <item type="id" name="fragmentLearn" />
     <item type="id" name="fragmentDiscover" />
     <item type="id" name="fragmentDownloads" />
+    <item type="id" name="fragmentDates" />
     <item type="id" name="fragmentProfile" />
 </resources>
