+++ core/src/main/java/org/openedx/core/config/AppLevelDatesConfig.kt
@@ -2,7 +2,7 @@ package org.openedx.core.config
 
 import com.google.gson.annotations.SerializedName
 
-data class DatesConfig(
+data class AppLevelDatesConfig(
     @SerializedName("ENABLED")
     val isEnabled: Boolean = true,
 )
