+++ core/src/main/java/org/openedx/core/config/ExperimentalFeaturesConfig.kt
@@ -5,4 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class ExperimentalFeaturesConfig(
     @SerializedName("APP_LEVEL_DOWNLOADS")
     val appLevelDownloadsConfig: AppLevelDownloadsConfig = AppLevelDownloadsConfig(),
+    @SerializedName("APP_LEVEL_DATES")
+    val appLevelDatesConfig: AppLevelDatesConfig = AppLevelDatesConfig(),
 )
