+++ core/src/main/java/org/openedx/core/data/model/CourseDatesResponse.kt
@@ -50,7 +50,9 @@ data class CourseDatesResponse(
             count = count,
             next = next,
             previous = previous,
-            results = results.mapNotNull { it.mapToDomain() }
+            results = results
+                .mapNotNull { it.mapToDomain() }
+                .sortedBy { it.dueDate }
         )
     }
 }
