+++ core/src/main/java/org/openedx/core/data/model/room/CourseDatesResponseEntity.kt
@@ -1,17 +1,55 @@
-package org.openedx.dates.data.storage
+package org.openedx.core.data.model.room
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import org.openedx.core.data.model.CourseDate
+import org.openedx.core.data.model.CourseDatesResponse
 import org.openedx.core.utils.TimeUtils
 import org.openedx.core.domain.model.CourseDate as DomainCourseDate
+import org.openedx.core.domain.model.CourseDatesResponse as DomainCourseDatesResponse
 
-@Entity(tableName = "course_date_table")
-data class CourseDateEntity(
+@Entity(tableName = "course_dates_response_table")
+data class CourseDatesResponseEntity(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo("course_date_id")
+    @ColumnInfo("course_date_response_id")
     val id: Int,
+    @ColumnInfo("course_date_response_count")
+    val count: Int,
+    @ColumnInfo("course_date_response_next")
+    val next: String?,
+    @ColumnInfo("course_date_response_previous")
+    val previous: String?,
+    @ColumnInfo("course_date_response_results")
+    val results: List<CourseDateDB>
+) {
+    fun mapToDomain(): DomainCourseDatesResponse {
+        return DomainCourseDatesResponse(
+            count = count,
+            next = next,
+            previous = previous,
+            results = results
+                .mapNotNull { it.mapToDomain() }
+                .sortedBy { it.dueDate }
+        )
+    }
+
+    companion object {
+        fun createFrom(courseDatesResponse: CourseDatesResponse): CourseDatesResponseEntity {
+            with(courseDatesResponse) {
+                return CourseDatesResponseEntity(
+                    id = 0,
+                    count = count,
+                    next = next,
+                    previous = previous,
+                    results = results.map { CourseDateDB.createFrom(it) }
+                )
+            }
+        }
+    }
+}
+
+data class CourseDateDB(
     @ColumnInfo("course_date_first_component_block_id")
     val firstComponentBlockId: String?,
     @ColumnInfo("course_date_courseId")
@@ -42,10 +80,9 @@ data class CourseDateEntity(
     }
 
     companion object {
-        fun createFrom(courseDate: CourseDate): CourseDateEntity {
+        fun createFrom(courseDate: CourseDate): CourseDateDB {
             with(courseDate) {
-                return CourseDateEntity(
-                    id = 0,
+                return CourseDateDB(
                     courseId = courseId,
                     firstComponentBlockId = firstComponentBlockId,
                     dueDate = dueDate,
