+++ dates/src/main/java/org/openedx/dates/data/repository/DatesRepository.kt
@@ -1,10 +1,10 @@
 package org.openedx.dates.data.repository
 
 import org.openedx.core.data.api.CourseApi
+import org.openedx.core.data.model.room.CourseDatesResponseEntity
 import org.openedx.core.data.storage.CorePreferences
 import org.openedx.core.domain.model.CourseDate
 import org.openedx.core.domain.model.CourseDatesResponse
-import org.openedx.dates.data.storage.CourseDateEntity
 import org.openedx.dates.data.storage.DatesDao
 
 class DatesRepository(
@@ -18,12 +18,22 @@ class DatesRepository(
         if (page == 1) {
             dao.clearCachedData()
         }
-        dao.insertCourseDateEntities(response.results.map { CourseDateEntity.createFrom(it) })
+        dao.insertCourseDateResponses(CourseDatesResponseEntity.createFrom(response))
         return response.mapToDomain()
     }
 
     suspend fun getUserDatesFromCache(): List<CourseDate> {
-        return dao.getCourseDateEntities().mapNotNull { it.mapToDomain() }
+        return dao.getCourseDateResponses()
+            .map { it.mapToDomain() }
+            .map { it.results }
+            .flatten()
+            .sortedBy { it.dueDate }
+    }
+
+    suspend fun preloadFirstPageCachedDates(): CourseDatesResponse? {
+        return dao.getCourseDateResponses()
+            .find { it.previous == null }
+            ?.mapToDomain()
     }
 
     suspend fun shiftDueDate() = api.shiftDueDate()
