+++ default_config/dev/config.yaml
@@ -31,9 +31,6 @@ PROGRAM:
 DASHBOARD:
   TYPE: 'gallery'
 
-APP_LEVEL_DATES:
-  ENABLED: true
-
 FIREBASE:
   ENABLED: false
   CLOUD_MESSAGING_ENABLED: false
