+++ default_config/stage/config.yaml
@@ -31,9 +31,6 @@ PROGRAM:
 DASHBOARD:
   TYPE: 'gallery'
 
-APP_LEVEL_DATES:
-  ENABLED: true
-
 FIREBASE:
   ENABLED: false
   CLOUD_MESSAGING_ENABLED: false
@@ -70,6 +67,8 @@ BRANCH:
 EXPERIMENTAL_FEATURES:
   APP_LEVEL_DOWNLOADS:
     ENABLED: false
+  APP_LEVEL_DATES:
+    ENABLED: true
 
 #Platform names
 PLATFORM_NAME: "OpenEdX"
