+++ downloads/src/main/java/org/openedx/downloads/presentation/download/DownloadsScreen.kt
@@ -78,7 +78,7 @@ import org.openedx.core.module.db.DownloadedState
 import org.openedx.core.module.db.DownloadedState.LOADING_COURSE_STRUCTURE
 import org.openedx.core.ui.HandleUIMessage
 import org.openedx.core.ui.IconText
-import org.openedx.core.ui.MainToolbar
+import org.openedx.core.ui.MainScreenToolbar
 import org.openedx.core.ui.OfflineModeDialog
 import org.openedx.core.ui.OpenEdXButton
 import org.openedx.core.ui.OpenEdXDropdownMenuItem
@@ -130,7 +130,7 @@ fun DownloadsScreen(
             .fillMaxSize(),
         backgroundColor = MaterialTheme.appColors.background,
         topBar = {
-            MainToolbar(
+            MainScreenToolbar(
                 modifier = Modifier
                     .statusBarsInset()
                     .displayCutoutForLandscape(),
